+++ frontend/components/article/Head.vue
@@ -1,17 +1,20 @@
 <template>
-    <NavigationButton  v-if="isArticlePage"  :data="button" :cls="['orange-btn']"/>
-
-    <section :class="[cls[0], {'article-section': !!isArticlePage}]">
-        <section :class="[{'blog-header': !isArticlePage}, {'ingress-header': isArticlePage}]">
-            <h1> {{ article.title }}</h1>
-            <p class="flex-wrap-row-align-items-center-justify-center article-metadata">
-                <span v-if="!!article.date" :class="cls[3]"> Publisert: <b><time :datetime="article.date.date">{{ article.date.date }}</time></b></span>
-                <NavigationAnchor v-for="(tag) in article.tags" :data="tag" :class="tag.cls"/>
-            </p>
-            <MDC :value="article.ingress" class="ingress-content" />
-            <NavigationNavMenu v-if="!isArticlePage && !!article.anchor" :data="article.anchor" :cls="['nav-bar', 'read-more']"
-        />
-        </section>
+    <NavigationButton  v-if="isArticlePage"  :data="button" :class="['orange-btn']"/>
+    <section :class="['flex-column-align-items-center', {'article-section': !!isArticlePage}]">
+        <Suspense>
+            <template #default>
+                <section :class="[{'blog-header': !isArticlePage}, {'ingress-header': isArticlePage}]">
+                    <h2>{{ article.title }}</h2>
+                    <p class="flex-wrap-row-align-items-center-justify-center article-metadata">
+                        <span v-if="!!article.date" :class="'meta-date'"> Publisert: <b><time :datetime="article.date.date">{{ article.date.date }}</time></b></span>
+                        <NavigationAnchor v-for="(tag) in article.tags" :data="tag" :class="tag.cls" />
+                    </p>
+                    <MDC :value="article.ingress" class="ingress-content" />
+                    <NavigationNavMenu v-if="!!article.anchor" :data="article.anchor" :class="['nav-bar', 'read-more']"/>
+                </section>
+            </template>
+            <template #fallback> <section class="loading">Laster innlegg til logger...</section> </template>
+        </Suspense>
     </section>
 </template>
 
@@ -23,9 +26,8 @@
     import type { ButtonItem } from '@/types/navigation';
 
     //  --- Props logic
-    const props = withDefaults(defineProps<HeaderProps>(), { cls: () => ['flex-column-align-items-center'], isArticlePage: () => false });
-    
-    const cls = props.cls;
+    const props = withDefaults(defineProps<HeaderProps>(), { isArticlePage: () => false });
+
     const article = computed(() => props.article);
     const isArticlePage = computed(() => props.isArticlePage);
 
@@ -37,4 +39,5 @@
     //console.log("Article Header Component - isNewsPage :", isNewsPage.value);
     //console.log("Article Header Component - isArticlePage :", isArticlePage.value);
     //console.log("Article Header Component - Article Anchor :", article.anchor);
+    //console.log("Article Header Component - Article Ingress :", article.value.tags);
 </script>