+++ frontend/components/timeline/Card.vue
@@ -4,11 +4,8 @@
 
         <section class="flex-wrap-row-align-items-center-justify-space-between">
             <section class="section">
-                <h3 v-if="!!data.organization.anchor">
-                    <NavigationAnchor :data="data.organization.anchor" />
-                </h3>
+                <h3 v-if="!!data.organization.anchor"> <NavigationAnchor :data="data.organization.anchor" /> </h3>
                 <h3 v-else>{{ data.organization.name }}</h3>
-
                 <h4 v-if="!!data.title"> {{ data.title }}</h4>
 
                 <section v-if="!!data.date"
@@ -30,32 +27,20 @@
                 <section v-if="!!data.techStack && data.techStack.length > 0"
                     :class="['tech-container']">
                     <h4>Teknologi(er) : </h4>
-                    <p :class="['flex-wrap-row-justify-space-evenly']">
-                        <span v-for="(tech, i) in data.techStack" :key="i">
-                            <MediaFigure v-if="tech" :data="tech"
-                                :cls="['tech-figure', 'tech-img']"
-                            />
-                        </span>
-                    </p>
+                    <section :class="['flex-wrap-row-justify-space-evenly']"> 
+                        <template v-for="(tech, i) in data.techStack" :key="i">
+                            <MediaFigure  v-if="tech"  :data="tech"  :cls="['tech-figure', 'tech-img']"  />
+                        </template>
+                    </section>
                 </section>
 
                 <section :class="cls[2]">
-                    <h3 v-if="!!data.location.anchor.href">
-                        <NavigationAnchor :data="data.location.anchor" />
-                    </h3>
-                    <h3 v-else>
-                        {{ data.location.name }}
-                    </h3>
-                    
-                    <h3 v-if="!!data.reference.anchor">
-                        <NavigationAnchor  :data="data.reference.anchor" />
-                    </h3>
+                    <h3 v-if="!!data.location.anchor.href"> <NavigationAnchor :data="data.location.anchor" /> </h3>
+                    <h3 v-else> {{ data.location.name }} </h3>
+                    <h3 v-if="!!data.reference.anchor"> <NavigationAnchor  :data="data.reference.anchor" /> </h3>
                 </section>
             </section>
-
-            <section class="section">
-                    <ContentRenderer v-if="data.body" :value="data.body" />
-            </section>
+            <section class="section"> <ContentRenderer v-if="data.body" :value="data.body" /> </section>
         </section>
 
     </section>