+++ frontend/pages/index.vue
@@ -1,32 +1,31 @@
 <template>
     <article class="article-wrapper flex-column">
-        <h2> Logger fra mine prosjekter </h2>
-        <section v-if="totalPages > 1" class="flex-wrap-row-align-items-center-justify-space-evenly">
-            <NavigationButton v-if="currentPage > 1" :data="lastPageBtn" :cls="['button', 'pagination-btn']"/>
+        <h2> Tekniske Logger </h2>
+        <section v-if="totalPages > 1" class="flex-wrap-row-align-items-center-justify-space-evenly pagination-container">
+            <NavigationButton v-if="currentPage > 1" :data="prevPage" :cls="['button', 'pagination-btn']"/>
                 <span> {{ currentPage }} / {{ totalPages }}</span>
-            <NavigationButton v-if="currentPage < totalPages" :data="nextPageBtn" :cls="['button', 'pagination-btn']"/>
+            <NavigationButton v-if="currentPage < totalPages" :data="nextPage" :cls="['button', 'pagination-btn']"/>
         </section>
 
         <section class="blog-section flex-wrap-row-align-items-center-justify-space-evenly">
                 <section v-for="post in mappedPosts" :key="post.id" class="blog-content">
                 <ArticleHead :article="post" />
             </section>
-
         </section>
     </article>
 
     <section :class="['flex-wrap-row-justify-space-evenly']">
 
-        <Timeline v-if="academicTimeline.length > 0"
+        <Timeline v-if="academicData.length > 0"
             title="Akademisk Tidslinje"
-            :data="academicTimeline"
+            :data="academicData"
             :cls = "['component-blue', 'timeline-container',
             'timeline-line', 'flex-wrap-row-justify-space-evenly', 'component-w-g-b']"
         />
 
-        <Timeline v-if="achievementsTimeline.length > 0"
+        <Timeline v-if="achievementData.length > 0"
             title="Prestasjonstidslinje"
-            :data="achievementsTimeline"
+            :data="achievementData"
             :cls = "['component-seagreen', 'timeline-container',
             'timeline-line', 'flex-wrap-row-justify-space-evenly', 'component-w-g-b']"
         />
@@ -44,22 +43,22 @@
     
     import { blogPagination } from '@/composables/pagination';
     import { mapTimeline } from '@/composables/maps/mapTimeline';
+    import { mapBlogData } from '~/composables/maps/mapBlogPost';
 
     import type { ButtonItem } from '~/types/navigation';
     import type { DevPostsCollectionItem, AcademicCollectionItem, AchievementsCollectionItem } from '@nuxt/content';
 
 
-    //  --- Component logic
-    const academicData = await fetchCollection<AcademicCollectionItem>('academic', 'academic-info');
-    const achievementData = await fetchCollection<AchievementsCollectionItem>('achievements', 'achievements-info');
 
-    const academicTimeline = computed(() => mapTimeline(academicData));
-    const achievementsTimeline = computed(() => mapTimeline(achievementData));
+    //  --- Component logic
+    const academicData = await fetchCollection<AcademicCollectionItem, ReturnType<typeof mapTimeline>>('academic', 'academic-info', mapTimeline);
+    const achievementData = await fetchCollection<AchievementsCollectionItem, ReturnType<typeof mapTimeline>>('achievements', 'achievements-info', mapTimeline);
+
 
         //  --- Conent logic
     const devPostPath = 'devPosts';
     const devPostCache = 'devPostCache';
-    const rawPosts = await fetchCollection<DevPostsCollectionItem>(devPostPath, devPostCache)
+    const rawPosts = await fetchCollection<DevPostsCollectionItem, ReturnType<typeof mapBlogData>>(devPostPath, devPostCache, mapBlogData);
     
     const n = 2; // Number of posts per page
     const currentPage: Ref<number> = ref(1);
@@ -67,8 +66,8 @@
     
     //  --- Pagination Logic
     const totalPages = computed(() => { if (rawPosts.value) {return Math.ceil(rawPosts.value.length / n); } return 0; });
-    const lastPageBtn = computed<ButtonItem>(() => { return { label: 'Forrige',  action: (): number => currentPage.value -- }; });
-    const nextPageBtn = computed<ButtonItem>(() =>  { return {label: 'Neste', action: ():number => { if (typeof currentPage.value === 'number') return currentPage.value++; else return 0;}};});
+    const prevPage = computed<ButtonItem>(() => { return { label: 'Forrige',  action: (): number => currentPage.value -- }; });
+    const nextPage = computed<ButtonItem>(() =>  { return {label: 'Neste', action: ():number => { if (typeof currentPage.value === 'number') return currentPage.value++; else return 0;}};});
 
     
     //  --- Debugging Logic