+++ frontend/pages/logs/records/[slug].vue
@@ -8,6 +8,7 @@
     import { mapBlogData } from '~/composables/maps/mapBlogPost';
 
     import type { DevPostsCollectionItem } from '@nuxt/content';
+import type { PostItem } from '~/types/documents';
     
 
      //  --- Route & slug logic
@@ -17,23 +18,22 @@
         //  --- Dev Data Logic
     const devPath = 'devPosts';
     const devCache = 'devCache';
-    const rawDev = await fetchCollection<DevPostsCollectionItem>(devPath, devCache)
+    const devPosts = await fetchCollection<DevPostsCollectionItem, ReturnType<typeof mapBlogData>>(devPath, devCache, mapBlogData);
 
     const personalPath = 'personalPosts';
     const personalCache = 'personalCache';
-    const rawPersonal = await fetchCollection<DevPostsCollectionItem>(personalPath, personalCache);
+    const personalPosts = await fetchCollection<DevPostsCollectionItem, ReturnType<typeof mapBlogData>>(personalPath, personalCache, mapBlogData);
     
     const article = computed(() => 
     {
         const currentSlug = String(slug);
 
-        const findBlog = (collection: DevPostsCollectionItem[] | undefined) => {
+        const findBlog = (collection: PostItem[] | undefined) => {
             if (!collection) return null;
-            const mapped = mapBlogData(collection);
-            return mapped?.find(blog => String(blog.path) === currentSlug) || null;
+            return collection.find(blog => String(blog.path) === currentSlug) || null;
         };
 
-        return findBlog(rawDev.value) ?? findBlog(rawPersonal.value);
+        return findBlog(devPosts.value) ?? findBlog(personalPosts.value);
 
     });