+++ frontend/pages/logs/tags/[slug].vue
@@ -1,6 +1,6 @@
 <template>
     <article class="article-wrapper flex-column">
-        <h2> Logger fra {{ slug }} Prosjektet</h2>
+        <h2> Logger fra {{ slug.charAt(0).toUpperCase() + String(route.params.slug).slice(1).replace(/-/g, ' ') }} Etiketten</h2>
         <section class="blog-section flex-wrap-row-align-items-center-justify-space-evenly">
                 <section v-for="post in article" :key="post.id" class="blog-content">
                 <ArticleHead :article="post" />
@@ -16,32 +16,30 @@
     import { fetchCollection } from '#imports';
     import { mapBlogData } from '~/composables/maps/mapBlogPost';
 
+    import type { PostItem } from '~/types/documents';
     import type { DevPostsCollectionItem } from '@nuxt/content';
 
 
      //  --- Route & slug logic
     const route = useRoute();
-    const slug = route.params.slug;
+    const slug = route.params.slug as string;
 
-        //  --- Dev Data Logic
     const devPath = 'devPosts';
     const devCache = 'devCache';
-    const rawDev = await fetchCollection<DevPostsCollectionItem>(devPath, devCache)
+    const devPosts = await fetchCollection<DevPostsCollectionItem, ReturnType<typeof mapBlogData>>(devPath, devCache, mapBlogData);
 
     const personalPath = 'personalPosts';
     const personalCache = 'personalCache';
-    const rawPersonal = await fetchCollection<DevPostsCollectionItem>(personalPath, personalCache);
-
+    const personalPosts = await fetchCollection<DevPostsCollectionItem, ReturnType<typeof mapBlogData>>(personalPath, personalCache, mapBlogData);
+    
     const article = computed(() => 
     { const currentSlug = String(slug);
 
-        const findBlog = (collection: DevPostsCollectionItem[]) => {
+        const findBlog = (collection: PostItem[] | undefined) => {
             if (!collection) return null;
-            console.log("Finding blog in collection: ", currentSlug);
-            const mapped = mapBlogData(collection);
-            return mapped?.filter(item => item.tags.some(tag => tag?.label === currentSlug));
+            return collection?.filter(item => item.isPublished && item.tags.some(tag => tag?.name === currentSlug));
         };
-        return findBlog(rawDev.value) ?? findBlog(rawPersonal.value);
+        return findBlog(devPosts.value) ?? findBlog(personalPosts.value);
 
     });