+++ frontend/stores/languageBytesStore.ts
@@ -1,5 +1,5 @@
+import { computed } from 'vue'
 import { defineStore } from 'pinia'
-import { ref, computed } from 'vue'
 
 export const useLanguageStore = defineStore('tech-language', () => {
 
@@ -8,19 +8,24 @@ export const useLanguageStore = defineStore('tech-language', () => {
   const allLanguages = computed(() => languages)
   const formattedLanguages = computed(() => {
     const object = sortByBytes();
-    const kb = 1024;
-    return Object.entries(object).filter(([_, bytes]) => bytes >= (kb*100)).map(([language, bytes]) => {
-      const kiloBytes: number = 1024;
-      const kb: number = bytes / kiloBytes;
 
-      return {
-        label: language.charAt(0).toUpperCase() + language.slice(1),
-        bytes: Number(kb.toFixed(2)),
-        percentage: Object.values(object).reduce((acc, val) => acc + val, 0) > 0 ? (kb / Object.values(object).reduce((acc, val) => acc + val, 0)) * 100 : 0
+    return Object.entries(object).map(([language, x]) => {
+      const bytes: number = 1024;
+      const kb: number = x / bytes;
+
+      return { 
+        type: 'KB', bytes: Number(kb.toFixed(2)), label: language.charAt(0).toUpperCase() + language.slice(1),
+        original: x,
+        percentage: Object.values(object).reduce((acc, val) => acc + val, 0) > 0 ? (kb / Object.values(object).reduce((acc, val) => acc + val, 0)) * 100 : 0.
+        
       }
-    })
+    }).filter((item) => item.bytes >= 100)
   });
 
+
+  function resetBytes(): void {
+    for (const key in languages) languages[key] = 0;
+  }
 
   function increment(key:string, value: number) : void {
     if (!key || typeof value !== 'number') return;
@@ -28,8 +33,8 @@ export const useLanguageStore = defineStore('tech-language', () => {
     if (!languages[key]) languages[key] = 0;
     
     languages[key] += value;
-    //console.log(`Updated ${key}: ${languages[key]} bytes`);
   }
+
   function sortByBytes(): Record<string, number> {
     const sorted = Object.entries(languages).sort((a, b) => b[1] - a[1]);
     return Object.fromEntries(sorted);
@@ -37,6 +42,7 @@ export const useLanguageStore = defineStore('tech-language', () => {
 
 
   return {
+    resetBytes,
     languages,
     increment,
     allLanguages,