+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+## [1.15.10](https://github.com/NexusWeave/webapp-Portfolio/compare/v1.15.9...v1.15.10) (2026-04-15)
+
+
+### Bug Fixes
+
+* **Config:** 🐛 Restricted middleware to CORS_ORIGINS and prevents any lazy-loading issues. ([92be8aa](https://github.com/NexusWeave/webapp-Portfolio/commit/92be8aa23e6de9114454908c207836c9ee508d89))
+
+## [1.15.9](https://github.com/NexusWeave/webapp-Portfolio/compare/v1.15.8...v1.15.9) (2026-04-15)
+
+
+### Bug Fixes
+
+* disable allow_credentials in CORS middleware as the page does not require cookies or login. ([3d7f79e](https://github.com/NexusWeave/webapp-Portfolio/commit/3d7f79ebd8bf51696d153fa6ca9a08797bd05675))
+
+## [1.15.8](https://github.com/NexusWeave/webapp-Portfolio/compare/v1.15.7...v1.15.8) (2026-04-14)
+
+
+### Bug Fixes
+
+* add import for computed. ([41773d0](https://github.com/NexusWeave/webapp-Portfolio/commit/41773d0b244e8163a2450f3527631dc094aca6d7))
+* Ensure blog filtering checks for published status. ([312aa2a](https://github.com/NexusWeave/webapp-Portfolio/commit/312aa2a97aaf67f96d943e0a654a9495fa8a292a))
+* Prevented unexpected behavior, by adding export for LanguageData. ([17de4e6](https://github.com/NexusWeave/webapp-Portfolio/commit/17de4e6ae7e2afcd52761cf0302cbd2fe7649e0b))
+* Removed unused component as it causing crashes. ([1ea331f](https://github.com/NexusWeave/webapp-Portfolio/commit/1ea331f4e88657fe22b8ee697c98afe2033bac40))
+
 ## [1.15.7](https://github.com/NexusWeave/webapp-Portfolio/compare/v1.15.6...v1.15.7) (2026-04-11)